requirements.txt
Django==1.5.1
GDAL==1.7.3
djangorestframework==2.3.3
vectorformats==0.1
